refactor(api): split medicineSchedules route into service and repository layers
The medicineSchedules route was a ~740-line file mixing Express HTTP glue,
Zod validation, business logic (adherence stats, today-summary aggregation,
IST handling, Redis summary-cache invalidation) and 13 raw Supabase queries.
This decouples it into the Controller-Service-Repository shape already used by
pharmacy.repository/service, with no behaviour change.

- repositories/medicineSchedule.repository.ts (new): one thin method per
  Supabase query, each moved verbatim so the exact query chains are preserved.
  Methods return the raw { data, error } result rather than throwing, because
  the service needs the distinct per-operation error messages the handlers used.
- services/medicineSchedule.service.ts (new): owns the Zod schemas + date/time
  validators, Redis cache read/write/invalidation, IST date handling, the
  frequency/times consistency check, adherence-stats calculation and
  today-summary aggregation. Expected failures are raised as a typed HttpError
  (status + message + optional Zod details); everything else bubbles up as an
  unexpected error.
- routes/medicineSchedules.ts: now pure HTTP glue — extract req params/body/
  query/user, call the service, map the result (or HttpError) to a response.
  741 -> 126 lines; zero Supabase or Redis calls remain in the route.

All 40 existing endpoint tests pass unchanged (they assert exact Supabase call
sequences and JSON shapes), so response contracts are identical. No DB schema
changes.

Closes #3775
